requirements.txt
# Flask e extensões
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.8.0
python-engineio==4.5.1
Werkzeug==2.3.7

# Manipulação de sessão e segurança
Flask-Session==0.5.0
itsdangerous==2.1.2

# Servidor WSGI
gunicorn==21.2.0
eventlet==0.33.3

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0  # Para manipulação de imagens (pode ser útil para jogos de desenho)
PyJWT==2.8.0    # Para tokens de autenticação

# Ferramentas de desenvolvimento (opcionais)
pytest==7.4.2   # Para testes unitários
black==23.9.1   # Para formatação de código

